+++ src/main/kotlin/io/github/graphglue/authorization/AuthorizationChecker.kt
@@ -1,17 +1,11 @@
-package io.github.graphglue.data.repositories
+package io.github.graphglue.authorization
 
-import io.github.graphglue.authorization.Permission
-import io.github.graphglue.data.execution.CypherConditionGenerator
 import io.github.graphglue.definition.NodeDefinitionCollection
 import io.github.graphglue.model.Node
 import org.neo4j.cypherdsl.core.Cypher
 import org.neo4j.cypherdsl.core.renderer.Renderer
 import org.springframework.data.neo4j.core.ReactiveNeo4jClient
-import org.springframework.data.neo4j.core.awaitFirstOrNull
-import org.springframework.stereotype.Component
 import reactor.core.publisher.Mono
-import reactor.core.publisher.toMono
-import reactor.kotlin.core.publisher.toMono
 
 /***
  * Check if a node allows a certain permission
